+++ CHANGELOG.md
@@ -9,10 +9,11 @@ This project adheres to [Semantic Versioning](https://semver.org/).
   * MAVEN SEP
   * MAVEN in situ key parameters
   * REACH Dosimeter
-  * DMSP SSUSI SDR-disk data
+  * DMSP SSUSI SDR-disk and SDR2-disk data
 * New Features
   * Allow files to be unzipped after download
-  * Added custom `concat_data` method to TIMED-GUVI data
+  * Added custom `concat_data` method to JHUAPL methods, for TIMED-GUVI and
+    DMSP-SSUSI data
   * Added cleaning to TIMED-GUVI SDR imaging data
 * Bug Fixes
   * Fix general clean routine to skip transformation matrices